Hang app-wide events off of cm.app instead of goog.global.
This has several effects worthy of celebration:
- JS tests no longer blow up catastrophically and consume gigabytes of memory.  Why? [*]
- The references to goog.global must have generated a lot of other unnecessary (though non-catastrophic) work in gjstest.stringify.  Eliminating this work means that JS tests now run >4x faster with this change (total time reduced from ~29s to ~6.5s).
- Test setup can make a new cm.app and test teardown can dispose it, helping to ensure that each test has a clean environment with no listeners lingering from preceding tests.

[*] The sordid details.

Listener tokens (from cm.events) include a reference to the event source.  And when we pass an opt_obj argument to cm.events.listen, it keeps the tokens in a special "referring listeners" property on opt_obj, so that it can detach the listeners when disposing the opt_obj.  Thus, when the source is goog.global, many of our views get reference chains like this:

view -> referring listeners -> listener token -> goog.global -> cm -> everything

Any attempt to gjstest.stringify the view object recursively stringifies all its referents; because gjstest.stringify is not very smart, it walks from all objects to their constructors; and because JavaScript thinks that the string representation of a function is the source code of the function, gjstest.stringify serializes the *entire source code of Crisis Map* (probably with many parts repeated multiple times).  This... is bad.

The specific trigger in this case was that something in the TabPanelView construction set up listeners on goog.global that had handlers in cm.AppState, and various tests (e.g. EditPresenterTest, PanelViewTest) would then pass in a cm.AppState as a mock expectation.  gjstest would then perform a comparison against the cm.AppState, the comparison would fail fast (plain pointer comparison), and the attempt to generate the explanation of the mismatch would then gjstest.stringify the cm.AppState, thereby stringifying goog.global and everything else in the universe.

This is a problem that must have existed for quite some time but for some reason was merely expensive but not disastrous, until something in the newer panel code triggered it more severely.

There were multiple factors contributing to the severity of the problem:
1. Objects hold references to goog.global via the listener token.
2. gjstest.stringify tries to recursively print out all referents too ambitiously.
2a. gjstest.stringify places no limit on recursion depth.
2b. gjstest.stringify tries to print out each object's constructors.
2c. gjstest.stringify tries to print out functions and methods by letting JS dump all their source code.
3. cm.TestBase.equals always calls gjstest.stringify twice to construct a Matcher, even if the Matcher fails or is never executed.

This CL addresses only #1; there are still opportunities for further improvement.  Arguably #2 is more of a root cause; it's reasonable to expect that merely referring to goog.global, or another large object, should not cause mysterious memory explosions in test runs.

Replacing gjstest.stringify with a no-op only brings the test time down another 15% or so, from 6.5 s to 5.5 s, so this CL realizes most of the time savings that could be gotten from optimizing gjstest.stringify.
-------------
Created by MOE: http://code.google.com/p/moe-java
MOE_MIGRATED_REVID=58114859