Activate opacity slider in GME imagery layers.
This is a little tricky since the opacity support is
only available in the bleeding edge experimental/
unstable API version (3.15) so this implementation is
careful to not assume that this API version is in use.

Also, since this careful code can be simplified a lot
once the 3.15 Maps API is public, I hid it with MOE.
That might not be the right decision but I decided to
make a guess and be conservative. I think when 3.15 is
the release API version the simpler version of this
implementation could be un-hidden.

The slider only appears if the config contains an
appropriate parameter naming the GME layer types
that can support opacity changes. So the caller
has to load the right bleeding-edge Maps API version
and then explicitly say "and turn it on for imagery
layers", which is what the [] changes in this
CL do. Otherwise it should be dormant so it doesn't
cause a useless slider in viewers using the stable
Maps API v3.14.
-------------
Created by MOE: http://code.google.com/p/moe-java
MOE_MIGRATED_REVID=61073602