Prevents unnecessary resize events from being emitted from the TabPanelView and consolidates tab content sizing into the TabView.
Notes:

We are getting some layout thrashing[1] due to triggering a resize event whenever anything changes in the tab panel. However, we only really need to go through the resizing routine when either (a) the viewport size changes or (b) the panel is expanded/collapsed in the "below" orientation (we need to explicitly set a size on the map so we have to measure the available space after the expand/collapse).

I was also able to eliminate some size recalculations when changing tabs by creating one instance of the toolbar shared by all of the tabs. The TabView can now keep the same max panel height as long as neither (a) or (b) above doesn't occur.


[1] See item #2: http://calendar.perfplanet.com/2013/the-runtime-performance-checklist/
-------------
Created by MOE: http://code.google.com/p/moe-java
MOE_MIGRATED_REVID=63330697